Create repo native query and controller method to get gameplays stats

+++ src/main/java/pl/wj/bgstat/boardgame/model/dto/BoardGameGameplayStatsDto.java
@@ -11,5 +11,5 @@ public interface BoardGameGameplayStatsDto {
     /**
      * Represents the gameplay average time in minutes of board game in given period
      */
-    int getAvgTimeOfGameplay();
+    double getAvgTimeOfGameplay();
 }
